pyproject.toml
[project]
name = "make_it_dense"
version = "0.1"
description = "CNN used to fuse sparse input data from a rotating LiDAR scan into TSDF volumes."
readme = "README.md"
python = "^3.6"
license = {file = "LICENSE.txt"}
authors = [{name = "Ignacio Martin Vizzo", email = "ignaciovizzo@gmail.com"}]
keywords = ["VDB", "SLAM", "LiDAR", "Odometry", "Localization", "Mapping" ]
dependencies = [
    "diskcache>=5.4.0",
    "easydict>=1.9",
    "numpy>=1.21",
    "open3d>=0.13",
    "pandas>=1.4",
    "pytorch_lightning>=1.6",
    "PyYAML<6.0",
    "tqdm>=4.64",
    "vdbfusion>=0.1.6",
    "typer>=0.4.1",
    # "torch",
    # pyopenvdb
    # vdb_to_numpy
]

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.black]
line-length = 100

[tool.pylint.format]
max-line-length = "100"
